+++ Runtime/Transitioners/Modules/PivotModule.cs
@@ -37,9 +37,10 @@ namespace HiddenAchievement.CrossguardUi.Modules
         /// <inheritdoc />
         public void ForceComponentRule(Transform component, IStyleModuleRule rule)
         {
-            if (rule is not PivotModuleRule pivotRule) return;
-            RectTransform rtComponent = (RectTransform)component;
-            rtComponent.pivot = pivotRule.Pivot;
+            // We can't do this because it triggers RectTransformDimensionsChange, which is not allowed in Awake.
+            // if (rule is not PivotModuleRule pivotRule) return;
+            // RectTransform rtComponent = (RectTransform)component;
+            // rtComponent.pivot = pivotRule.Pivot;
         }
         
         /// <inheritdoc />